+++ services/autoscaling/tests/unit/test_autoscaling_cli_extended.py
@@ -14,7 +14,6 @@ class TestAutoScaling(unittest.TestCase):
         result = util.invoke_command(['autoscaling', 'policy', 'create'])
         assert 'Error: Missing option(s)' in result.output
         assert 'auto-scaling-configuration-id' in result.output
-        assert 'capacity' in result.output
         assert 'policy-type' in result.output
         result = util.invoke_command(['autoscaling', 'policy', 'create', '--auto-scaling-configuration-id', '1', '--capacity', '1', '--policy-type', 'scheduled'])
         assert 'UsageError: If Parameter --policy-type is scheduled, then Parameter --execution-schedule must be provided' in result.output